pyproject.toml
[project]
name = "aicodereview"
version = "0.1.0"
description = ""
authors = [
    {name = "Dodgemaster1",email = "mykyta.manaiev@gmail.com"}
]
readme = "README.md"
requires-python = " >=3.12"
dependencies = [
    "fastapi (>=0.115.6,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "pytest (>=8.3.4,<9.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "google-generativeai (>=0.8.3,<0.9.0)",
    "redis (>=5.2.1,<6.0.0)",
    "pytest-asyncio (>=0.25.2,<0.26.0)",
]


[tool.poetry]
name = "aicodereview"
version = "0.1.0"
description = ""
authors = ["Dodgemaster1 <mykyta.manaiev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
fastapi = ">=0.115.6,<0.116.0"
uvicorn = ">=0.34.0,<0.35.0"
pytest = ">=8.3.4,<9.0.0"
httpx = ">=0.28.1,<0.29.0"
google-generativeai = ">=0.8.3,<0.9.0"
redis = ">=5.2.1,<6.0.0"
pytest-asyncio = ">=0.25.2,<0.26.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
